Search for mkl_rt in the loaded libraries of the process
Instead of hardcoded library names, search for mkl_rt in the loaded
libraries of the python process. Mkl is loaded by "import mkl" and can
therefore be found when searching the loaded libraries. This is
platform-agnostic and more robust. A small change of the dll name on
windows from "mkl_rt.dll" to "mkl_rt.1.dll" broke pypardiso before.

Fixes #12